Add route handlers for book and recommendation services; implement reading patterns analysis
