+++ routes/__init__.py
@@ -1,0 +1,5 @@
+from .analysis_routes import analysis_bp
+from .book_routes import book_bp
+from .recommendation_routes import recommendation_bp
+
+__all__ = ['analysis_bp', 'book_bp', 'recommendation_bp']
